% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biggerthantwo.R
\name{biggerthantwo}
\alias{biggerthantwo}
\title{biggerthantwo}
\usage{
biggerthantwo(x)
}
\arguments{
\item{x}{A real number}
}
\value{
A string telling if "Bigger", "Smaller" or "Equal" to 2
}
\description{
The functions checks if the input value is bigger, smaller or equal to 2.
}
\examples{
bignumber <- biggerthantwo(1000)
bignumber
"Bigger"

}
