% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ismultiple13.R
\name{ismultiple13}
\alias{ismultiple13}
\title{ismultiple13}
\usage{
ismultiple13(x)
}
\arguments{
\item{x}{Real Value}
}
\value{
String expressing if the input value is a multiple of 13 or the residual.
}
\description{
The functions checks if the input value is a multiple of 13.
If not it returns the residual.
}
\examples{
x <- 15
multiple13(x)
Not Multiple. The residual is: 2

}
